注文のエラーチェックを厳密にした

+++ src/main/java/red/man10/UserData.java
@@ -53,7 +53,7 @@ public class UserData {
             last = his.get(1);
         }
 
-        String ret = "§f§l残高:"+Utility.getPriceString(today.bal) + "§f§lアイテムバンク評価額:"+Utility.getPriceString(today.estimated_value) +" §f§lアイテム個数:"+Utility.getItemString(today.total_amount);
+        String ret = "§f§l残高:"+Utility.getPriceString(today.bal) + "§f§lアイテムバンク評価額:"+Utility.getPriceString(today.estimated_value) +" §f§lアイテム個数:"+Utility.getItemString((int)today.total_amount);
 
 
         return ret;
